Fixed Tile Map, Created Template, & ReadMe
I removed the child of PaintLayer1 that was baked into the prefab as to not have a permanent hex at the origin of the map. I also moved the Directional Light far far away to remove any issues it causes with the ray cast to the tile map. I saved all of these changes and labeled this new scene as a template for our level designers to simply copy, rename, and get started. I also created a README with detailed instructions on how to get started, details on what problems they may run into, and how to export the map as an FBX.